fix stories, refactor
